+++ TEAM_1/develop/src/error/syntax_error.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   syntax_error.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kilee <kilee@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/11 15:32:35 by kilee             #+#    #+#             */
+/*   Updated: 2021/03/11 15:34:48 by kilee            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-t_bool	have_syntax_error(char *stdin_buf)
+t_bool		have_syntax_error(char *stdin_buf)
 {
 	if (too_many_semicolon(stdin_buf)
 		|| too_many_right_redirection(stdin_buf)
@@ -19,7 +31,7 @@ void		syntax_error(void)
 	g_exit_code = 258;
 }
 
-t_bool	too_many_semicolon(char *stdin_buf)
+t_bool		too_many_semicolon(char *stdin_buf)
 {
 	int		count;
 
@@ -37,7 +49,7 @@ t_bool	too_many_semicolon(char *stdin_buf)
 	return (FALSE);
 }
 
-t_bool	too_many_right_redirection(char *stdin_buf)
+t_bool		too_many_right_redirection(char *stdin_buf)
 {
 	int		count;
 
@@ -55,7 +67,7 @@ t_bool	too_many_right_redirection(char *stdin_buf)
 	return (FALSE);
 }
 
-t_bool	too_many_left_redirection(char *stdin_buf)
+t_bool		too_many_left_redirection(char *stdin_buf)
 {
 	int		count;
 
@@ -73,7 +85,7 @@ t_bool	too_many_left_redirection(char *stdin_buf)
 	return (FALSE);
 }
 
-t_bool	too_many_pipe(char *stdin_buf)
+t_bool		too_many_pipe(char *stdin_buf)
 {
 	int		count;
