+++ TEAM_1/develop/src/exec/exec_command.c
@@ -6,7 +6,7 @@
 /*   By: kilee <kilee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/10 13:41:26 by hson              #+#    #+#             */
-/*   Updated: 2021/03/11 15:03:55 by kilee            ###   ########.fr       */
+/*   Updated: 2021/03/11 16:21:43 by kilee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ void	exec_command(void)
 	cmd = g_cmd;
 	while (cmd)
 	{
-		// test_make_cmd_str_to_tokens();
+		test_make_cmd_str_to_tokens();
 		substitute_command(cmd);
 		// test_make_cmd_str_to_tokens();
 		substitute_redir(cmd);
